+++ tests/unit/dependency_checker/test_helm_dependency_checker.py
@@ -1,8 +1,11 @@
 """Tests for the Helm dependency checker."""
 
-from typing import List
+from types import FunctionType
+from typing import Dict, List
+from urllib.parse import parse_qs, urlparse
 
 import pytest
+import semver
 
 from ska_mid_itf_engineering_tools.dependency_checker import helm_dependency_checker
 from ska_mid_itf_engineering_tools.dependency_checker.types import Dependency
@@ -107,3 +110,176 @@ def test_parse_helm_invalid(helm_output_invalid: str):
     with pytest.raises(ValueError):
         dc = helm_dependency_checker.HelmDependencyChecker()
         dc.parse_helm_dependencies(helm_output_invalid)
+
+
+@pytest.fixture(name="helm_search_single_response")
+def fixture_helm_search_single_response() -> Dict:
+    """
+    Return a single Helm search response.
+
+    :return: The response.
+    :rtype: Dict
+    """
+    return {
+        "items": [
+            {
+                "name": "ska-tmc-mid",
+                "version": "0.4.1",
+            },
+            {
+                "name": "ska-tmc-mid",
+                "version": "0.6.0",
+            },
+        ],
+        "continuationToken": None,
+    }
+
+
+@pytest.fixture(name="helm_search_multiple_responses")
+def fixture_helm_search_multiple_responses() -> Dict:
+    """
+    Return multiple Helm search responses.
+
+    :return: The responses, with their continuation tokens as keys.
+    :rtype: Dict
+    """
+    return {
+        "none": {
+            "items": [
+                {
+                    "name": "ska-tmc-mid",
+                    "version": "0.4.1",
+                },
+                {
+                    "name": "ska-tmc-mid",
+                    "version": "0.6.0",
+                },
+            ],
+            "continuationToken": "first",
+        },
+        "first": {
+            "items": [
+                {
+                    "name": "ska-tmc-mid",
+                    "version": "0.3.1",
+                },
+                {
+                    "name": "ska-tmc-mid",
+                    "version": "0.8.0",
+                },
+            ],
+            "continuationToken": "second",
+        },
+        "second": {
+            "items": [
+                {
+                    "name": "ska-tmc-mid",
+                    "version": "0.1.1",
+                },
+                {
+                    "name": "ska-tmc-mid",
+                    "version": "0.9.0",
+                },
+            ],
+            "continuationToken": None,
+        },
+    }
+
+
+def helm_multiple_response_callback(data: Dict) -> FunctionType:
+    """
+    Create a mock response callback for processing multiple Helm responses.
+
+    :param data: The multiple Helm search responses dict to use.
+    :type data: Dict
+    :return: The callback
+    :rtype: FunctionType
+    """
+
+    def callback(request, context):
+        context.status_code = 200
+        o = urlparse(request.url)
+        query = parse_qs(o.query)
+        print(query)
+        token = query.get("continuationToken", None)
+        if token is None:
+            token = "none"
+        elif len(token) > 0:
+            token = token[0]
+        return data[token]
+
+    return callback
+
+
+def test_find_latest_chart_version_single_response_same_version(
+    requests_mock, helm_search_single_response: Dict
+):
+    """
+    Test finding the latest version of a chart when there is a single response and no new version.
+
+    :param requests_mock: requests_mock object
+    :param helm_search_single_response: Helm single response fixture
+    :type helm_search_single_response: Dict
+    """
+    requests_mock.get(
+        "https://artefact.skao.int/service/rest/v1/search", json=helm_search_single_response
+    )
+    dc = helm_dependency_checker.HelmDependencyChecker()
+    current = Dependency("ska-tmc-mid", "0.6.0", "0.6.0")
+    latest = dc.find_latest_chart_version(current)
+    assert latest == current.project_version
+
+
+def test_find_latest_chart_version_single_response_new_version(
+    requests_mock, helm_search_single_response: Dict
+):
+    """
+    Test finding the latest version of a chart when there is a single response and a new version.
+
+    :param requests_mock: requests_mock object
+    :param helm_search_single_response: Helm single response fixture
+    :type helm_search_single_response: Dict
+    """
+    requests_mock.get(
+        "https://artefact.skao.int/service/rest/v1/search", json=helm_search_single_response
+    )
+    dc = helm_dependency_checker.HelmDependencyChecker()
+    current = Dependency("ska-tmc-mid", "0.4.0", "0.4.0")
+    latest = dc.find_latest_chart_version(current)
+    assert latest == semver.Version(0, 6, 0)
+
+
+def test_find_latest_chart_version_multiple_responses_new_version(
+    requests_mock, helm_search_multiple_responses: Dict
+):
+    """
+    Test finding the latest version of a chart when there are multiple responses and a new version.
+
+    :param requests_mock: requests_mock object
+    :param helm_search_multiple_responses:  Helm multiple responses fixture
+    :type helm_search_multiple_responses: Dict
+    """
+    callback = helm_multiple_response_callback(helm_search_multiple_responses)
+    requests_mock.get("https://artefact.skao.int/service/rest/v1/search", json=callback)
+    dc = helm_dependency_checker.HelmDependencyChecker()
+    current = Dependency("ska-tmc-mid", "0.4.0", "0.4.0")
+    latest = dc.find_latest_chart_version(current)
+    assert latest == semver.Version(0, 9, 0)
+
+
+def test_find_latest_chart_version_multiple_responses_same_version(
+    requests_mock, helm_search_multiple_responses: Dict
+):
+    """
+    Find the latest version of a chart when there are multiple responses and no new version.
+
+    :param requests_mock: requests_mock object
+    :param helm_search_multiple_responses:  Helm multiple responses fixture
+    :type helm_search_multiple_responses: Dict
+    """
+    callback = helm_multiple_response_callback(helm_search_multiple_responses)
+    requests_mock.get("https://artefact.skao.int/service/rest/v1/search", json=callback)
+    dc = helm_dependency_checker.HelmDependencyChecker()
+    current = Dependency("ska-tmc-mid", "0.9.0", "0.9.0")
+    latest = dc.find_latest_chart_version(current)
+    assert latest == semver.Version(0, 9, 0)